+++ temba/channels/tests.py
@@ -1011,14 +1011,27 @@ class ChannelCRUDLTest(TembaTest, CRUDLTestMixin):
 
         # fields shown depend on scheme and role
         self.assertUpdateFetch(
-            android_url, [self.editor, self.admin], form_fields={"name": "My Android", "allow_international": False}
+            android_url,
+            [self.editor, self.admin],
+            form_fields={
+                "name": "My Android",
+                "is_enabled": True,
+                "allow_international": False,
+            },
         )
         self.assertUpdateFetch(
             vonage_url,
             [self.editor, self.admin],
-            form_fields={"name": "My Vonage", "allow_international": False, "machine_detection": False},
+            form_fields={
+                "name": "My Vonage",
+                "is_enabled": True,
+                "allow_international": False,
+                "machine_detection": False,
+            },
         )
-        self.assertUpdateFetch(telegram_url, [self.editor, self.admin], form_fields={"name": "My Telegram"})
+        self.assertUpdateFetch(
+            telegram_url, [self.editor, self.admin], form_fields={"name": "My Telegram", "is_enabled": True}
+        )
 
         # name can't be empty
         self.assertUpdateSubmit(
@@ -1033,7 +1046,12 @@ class ChannelCRUDLTest(TembaTest, CRUDLTestMixin):
         self.assertUpdateSubmit(
             vonage_url,
             self.admin,
-            {"name": "Updated Name", "allow_international": True, "machine_detection": True},
+            {
+                "name": "Updated Name",
+                "is_enabled": True,
+                "allow_international": True,
+                "machine_detection": True,
+            },
         )
 
         vonage_channel.refresh_from_db()
@@ -1045,14 +1063,19 @@ class ChannelCRUDLTest(TembaTest, CRUDLTestMixin):
         self.assertUpdateFetch(
             vonage_url,
             [self.editor, self.admin],
-            form_fields={"name": "Updated Name", "allow_international": True, "machine_detection": True},
+            form_fields={
+                "name": "Updated Name",
+                "is_enabled": True,
+                "allow_international": True,
+                "machine_detection": True,
+            },
         )
 
         # staff users see extra log policy field
         self.assertUpdateFetch(
             vonage_url,
             [self.customer_support],
-            form_fields=["name", "log_policy", "allow_international", "machine_detection"],
+            form_fields=["name", "is_enabled", "log_policy", "allow_international", "machine_detection"],
             choose_org=self.org,
         )