+++ temba/channels/types/twilio_messaging_service/tests.py
@@ -118,7 +118,7 @@ class TwilioMessagingServiceTypeTest(TembaTest):
         self.login(self.admin)
         response = self.client.get(update_url)
         self.assertEqual(
-            ["name", "allow_international", "account_sid", "auth_token", "loc"],
+            ["name", "is_enabled", "allow_international", "account_sid", "auth_token", "loc"],
             list(response.context["form"].fields.keys()),
         )